feat: 支持内联事件处理器在子应用作用域中执行
- 添加 getWujieWindow 辅助函数，用于在 ShadowDOM 中动态获取子应用 window
- 在 processTpl 中编译内联事件处理器，包裹 with(__getWujieWindow__(this)){}
- 解决传统页面中 onclick="greet(...)" 无法访问子应用函数的问题
- 添加相应的类型定义

问题背景：
wujie 采用 ShadowDOM + iframe 双沙箱架构，导致内联事件处理器在主应用
ShadowDOM 中执行时，无法访问 iframe 闭包中定义的函数。

解决方案：
在 HTML 解析阶段，将 onclick="..." 编译为
onclick="with(__getWujieWindow__(this)){ ... }"，
通过 DOM 查询动态获取子应用的 proxyWindow，实现作用域绑定。
